Trying to eliminate type ambiguity for compilers
compilers may have trouble generating the correct (optimized) prolog code when calling variadic functions and/or with template types that require variable/unpredictable local stack memory allocation; this attempts to give compilers more clues and explicitly allocates and frees heap memory for the temporary (swapped) vector list handle so that we (might) instruct the compiler to do that stack memory allocation first even when optimizing. the intent is for there to be only the subscriptions pointer and the handlers &handle as stack variables, so the compiler can do whatever it does for variadic function parameters plus these two stack variables and be done.